==> 02-data.Rmd <==
# Data sources

```{r}

covid_MX <- read.csv("Data/COVID_mex.csv", stringsAsFactors = F)
covid_US <- read.csv("Data/COVID_usa.csv", stringsAsFactors = F)

inflation_MX <- read.csv("Data/Inflation_MX_2019_2021.csv", stringsAsFactors = F)
inflation_US <- read.csv("Data/Inflation_US_2012_2021.csv", stringsAsFactors = F)

GDP_MX <- read.csv("Data/GDP_MX_Clean.csv", stringsAsFactors = F)
GDP_US <- read.csv("Data/GDP_US_Clean.csv", stringsAsFactors = F)
```

The data that we will use in the project can be categorized in three verticals:

* **Health data**: Consisting of positive COVID cases, deaths due to COVID, vaccination and policy implementation.

* **Macroeconomic data**: Consisting of economic growth by sector and the inflation rate.

* **Financial data**: Consisting of market indices.

## Health data

For this category of variables, team member Ulises Hernandez located a very well polished R package [COVID19](https://covid19datahub.io/)^[Guidotti, E., Ardia, D., (2020), "COVID-19 Data Hub", Journal of Open Source Software 5(51):2376, doi: 10.21105/joss.02376.] of data related to the COVID-19 pandemic. This repository was initiated by Emanuele Guidotti (University of Neuchâtel), leveraged by David Ardia (HEC Montréal), enhanced by the open source community, and is currently maintained by Emanuele Guidotti.

The initiative has aggregated data of many countries, but we restricted our use to only data corresponding to the US and Mexico. The variables in the data set can classified in three broad categories:

### Epidemiological variables

This group of variables includes the following variables:

* `confirmed`:	Cumulative number of confirmed cases.
* `deaths`:	Cumulative number of deaths.
* `recovered`:	Cumulative number of patients released from hospitals or reported recovered.
* `tests`:	Cumulative number of tests.
* `vaccines`:	Cumulative number of total doses administered.
* `people_vaccinated`:	Cumulative number of people who received at least one vaccine dose.
* `people_fully_vaccinated`:	Cumulative number of people who received all doses prescribed by the vaccination protocol.
* `hosp`:	Number of hospitalized patients on date.
* `icu`:	Number of hospitalized patients in intensive therapy on date.
* `vent`:	Number of patients requiring invasive ventilation on date.
* `population`: Total population.

### Policy measures

According to the documentation of the data hub, policy measures are provided by [Oxford Covid-19 Government Response Tracker](https://github.com/OxCGRT/covid-policy-tracker). As one can expect with this kind of data, policies not only vary by country, but even by region within given country. For this reason, the tracker uses flag variables to denote this limited scope of different measures. That is, the indicators have a flag for whether they are”targeted" to a specific geographical region (flag=0) or whether they are a “general” policy that is applied across the whole country/territory (flag=1). The compiled variables are the following:

* `school_closing`
* `workplace_closing`
* `cancel_events`
* `gatherings_restrictions`
* `transport_closing`
* `stay_home_restrictions`
* `internal_movement_restrictions`
* `international_movement_restrictions`
* `information_campaigns`
* `testing_policy`
* `contact_tracing`
* `facial_coverings`
* `vaccination_policy`
* `elderly_people_protection`
* `government_response_index`
* `stringency_index`
* `containment_health_index`
* `economic_support_index`

### Administrative areas

Finally, the data hub also compiles data on the administrative area:

* `administrative_area_level`: Level of the administrative area: 1 for countries; 2 for states, regions, cantons, or local * * `equivalent; 3 for cities, municipalities, or local equivalent.
* `administrative_area_level_1`:	Name of the administrative area of level 1.
* `administrative_area_level_2`:	Name of the administrative area of level 2.
* `administrative_area_level_3`:	Name of the administrative area of level 3.

For a more detailed explanation  of the variables just mentioned, the [documentation page](https://covid19datahub.io/articles/docs.html) of the data hub is an excellent resource for people interested in using this data set.

## Macroeconomic data

### Economic Growth

Team member Abel Perez consulted economic growth data for Mexico from https://www.inegi.org.mx/app/indicadores/?tm=0 , in specific, he consulted the monthly approximation to GDP (which is quarterly) known as the Global Economic Activity Indicator (IGAE in spanish). These data also present information about the production in different economic activities know as primary, secondary and third activities. The data was preliminary cleaned in Excel sheets (deleting comments on top of the data) and processed as presented in Chapter 3.

The data spans from January 2019 to September 2021 (33 observations), and after cleansing it contains the annual percentage change on economic activity for total, primary, secondary and third activities.

```{r}
str(GDP_MX)
```

For the US, he consulted the monthly GDP from https://ihsmarkit.com/products/us-monthly-gdp-index.html. This data do not include detail on the production of primary, secondary and third activities. Furthermore the definition of such activities changes between Mexico and the US. To tackle this problem, he consulted Mexico's aggregation of economic activities from https://www.banxico.org.mx/SieInternet/consultarDirectorioInternetAction.do?sector=2&accion=consultarCuadro&idCuadro=CR199&locale=es. He then consulted the proportion of contribution of sub-economic activities in the US from https://apps.bea.gov/iTable/iTable.cfm?reqid=150&step=2&isuri=1&categories=gdpxind and aggregated the activities based on their definition in Mexico to compute weights for primary, secondary and third activities in the US based on Mexico's definition. Finally, we simply multiplied such weights to the total GDP in the US to obtain the production for activities and then compute the annual percentage change (all this process is documented in the code of Chapter 3).

The data for US spans from January 2019 to September 2021 (33 observations), and has the same variables as in Mexico's data

```{r}
str(GDP_US)
```

### Inflation

Team member Abel Perez consulted the data for inflation in Mexico from the Central Bank of Mexico's System of Economic Inormation (https://www.banxico.org.mx/SieInternet/consultarDirectorioInternetAction.do?accion=consultarCuadro&idCuadro=CP154&locale=es). The monthly Index for Consumer Prices was downloaded and inflation was computed in an auxiliar Excel sheet as the annual precentage change of the Index.

The data of the Index of Cosumer Prices for the US was consulted in a similar manner from US Bureau of Labor Statistics (https://data.bls.gov/cgi-bin/surveymost) and inflation was computed in the same way.

The data for Mexican inflation we consulted spans from January 2019 to October 2021, consisting of 34 observations:

```{r}
str(inflation_MX)
```

The data for US inflation spans from January 2012 to October 2021 (we will only focus on the intersection of time windows with Mexico), consisting of 118 observations:

```{r}
str(inflation_US)
```


## Financial data

For this category of variables, team member Gilberto Garcia used the R package [tidyquant](https://cloud.r-project.org/web/packages/tidyquant/index.html), which connects to number of data sources including Yahoo Finance, Morningstar, Quandl, FRED Economic Data and more, to download the data in a format compatible with tidyverse. The code used to download this data can be found [here](https://github.com/EDAV-Section2-FP34/covidUSMexico/blob/main/Data//financial data downloader.R).

In the case of the United States, given that there are plenty of well maintained financial data sources, we decided to pick the S&P 500 index as a representative proxy of the health of the American financial market, and also obtained data on the 11 sectors making up the index:

1. Materials
2. Health Care
3. Consumer Staples
4. Consumer Discretionary
5. Energy
6. Financials
7. Industrials
8. Technology
9. Utilities
10. Real State
11. Communication Services

In the case of Mexico there was a natural limitation, which is that the size of the Mexican financial market is much smaller in terms of number of companies and market capitalization. As a reference, the Índice de Precios y Cotizaciones (IPC), the only Mexican stock index, is composed of around 35 publicly traded companies. We gathered data on this index, but were unable to obtained sectorized data because of this limitation.

The API we used downloaded these data as a data frame with 9,671 observations and 8 columns. The columns correspond to the following variables: symbol (ticker), date, open price of trading day , highest price of trading day , lowest price of trading day , closing price of trading, volume of trading, and finally and most importantly the adjusted price of the index, which is the price adjusted by corporate actions and is the relevant price we have to focus.

Finally, for the purposes of making value comparison across countries, we downloaded data on the exchange rate between the Mexican Peso and the US dollar.
