==> 03-cleaning.Rmd <==
# Data transformation

Depending on the type of variables, different data transformations were applied in order to make the proper analysis. As is natural with time-series data, many of the transformations had to do with a way of dealing with trends or change in the longitudinal, although in some instances this trending behaviour was desired in order to make proper comparisons.

## Health variables

In this category of variables, the main purpose of the transformations was to provide a way to make comparison of daily cases and deaths related to COVID-19, and also vaccination rates when this information became available. The methodology we followed was to take the ratio of the Mexican target metric over the American target metric. For example, in the case of daily COVID-19 confirmed cases, we created a new time series that was calculated as by dividing the daily count of COVID-19 cases in Mexico divided by the corresponding daily count of COVID-19 cases in the United States. Therefore, a number greater that 1 for this index means that coronavirus cases in Mexico increased more rapidly than cases in the United States, and vice versa for an index value less than 1. The code we used for transforming this data can be found [here](https://github.com/EDAV-Section2-FP34/covidUSMexico/blob/main/datos_series.R).

## Economic variables

The main challenge in this category of data was to match the different definitions of economic activity in the Mexican and in the US. Government agencies in Mexico and the US follow classifications of economic enterprises and data, and hence, the bulk of the transformations we used here focused on making an appropriate matching of levels in the two countries.

## Financial variables

Finally, financial variables data required the standardization of the price (levels) of the different indices that were gathered, in a way that would clearly show the sectors and indices that outperformed the index average, as well as to give a clear picture of abrupt movement of financial markets when pandemic begun. In addition, given that this data was downloaded through an API, the data required the usual pivot longer transformations in order to be arranged in a desired format. The code we used to download these data and make the initial set of transformations can be found [here](https://github.com/EDAV-Section2-FP34/covidUSMexico/blob/main/Data/financial%20data.R).

```{r}
library(dplyr)
library(tidyr)

# GDP for Mexico. Monthly index for total, first, second and third economic activities
GDP_MX <- read.csv("Data/IGAE_Brute.csv", stringsAsFactors = F)
GDP_MX$Periodo <- 100*as.numeric(substring(GDP_MX$Periodo, 1, 4)) + 
  as.numeric(substring(GDP_MX$Periodo, 6, 7))

# Function to compute year percent changes
#
# Inputs: x, a numeric vector
#
# Output: y, a vector defined as NA for the first k elements, and x_t/x_t-k -1 for the
#         rest
pch <- function(x, k=12){
  y <- x[-(1:k)]/x[-( (length(x)-k+1):length(x) )] - 1
  return( c(rep(NA, k), 100*y) ) 
}

# Year percentage changes
GDP_MX[, -1] <- sapply(GDP_MX[, -1], pch)
GDP_MX <- GDP_MX %>% filter(!is.na(Total))

# Saving
GDP_MX %>% write.csv("Data/GDP_MX_Clean.csv", row.names = F)


# GDP for US. Monthly level (total) 
GDP_US <- read.csv("Data/GDP_US_Brute.csv", stringsAsFactors = F)
aux <- substring(GDP_US$Month, 8,10)
GDP_US$M <- ifelse(aux=="Jan", 1,
                   ifelse(aux=="Feb", 2,
                          ifelse(aux=="Mar", 3,
                                 ifelse(aux=="Apr", 4,
                                        ifelse(aux=="May", 5,
                                               ifelse(aux=="Jun", 6,
                                                      ifelse(aux=="Jul", 7,
                                                             ifelse(aux=="Aug", 8,
                                        ifelse(aux=="Sep", 9,
                                               ifelse(aux=="Oct", 10,
                                                      ifelse(aux=="Nov", 11, 12 )))))))) )  ) )
GDP_US$OBS <- 100*as.numeric(substring(GDP_US$Month, 1, 4)) + GDP_US$M


# Proportion of GDP of first, second and third economic activities, following Mexico's
# definition in https://www.banxico.org.mx/SieInternet/consultarDirectorioInternetAction.do?sector=2&accion=consultarCuadro&idCuadro=CR199&locale=es
#
# Obtained from https://apps.bea.gov/iTable/iTable.cfm?reqid=150&step=2&isuri=1&categories=gdpxind
GDP_US_Activities <- read.csv("Data/Activity_Share_2.csv", stringsAsFactors = F)
GDP_US_Activities <- GDP_US_Activities %>% pivot_longer(cols = !c(Activity, Sector),
                                                        names_to="OBS", values_to="w")
GDP_US_Activities$OBS <- as.numeric(gsub("X", "", GDP_US_Activities$OBS))
GDP_US_Activities <- GDP_US_Activities %>% group_by(OBS, Sector) %>% 
  summarise(w=sum(w)) %>% ungroup() %>% arrange(Sector, OBS)

# GDP + Economic Activity
GDP_US$Quarter <- ifelse(GDP_US$M / 3 <= 1, 1,
                         ifelse(GDP_US$M / 3 <= 2, 4, 
                                ifelse(GDP_US$M / 3 <= 3, 7, 10)  ) )
GDP_US$OQuarter <- 100*as.numeric(substring(GDP_US$Month, 1, 4)) + GDP_US$Quarter
GDP_US$OQuarter[GDP_US$OQuarter==202107] <- 202104
w1 <- GDP_US_Activities %>% filter(Sector==1) %>% dplyr::select(!Sector)
w2 <- GDP_US_Activities %>% filter(Sector==2) %>% dplyr::select(!Sector)
w3 <- GDP_US_Activities %>% filter(Sector==3) %>% dplyr::select(!Sector)
GDP_US <- GDP_US %>% filter(OBS>=201801) %>% left_join(w1, by=c('OQuarter'='OBS')) %>% 
   left_join(w2, by=c('OQuarter'='OBS')) %>% left_join(w3, by=c('OQuarter'='OBS'))
GDP_US <- GDP_US %>% mutate(Total=Monthly.Nominal.GDP.Index,
                            Primarias=Monthly.Nominal.GDP.Index*w.x,
                            Secundarias=Monthly.Nominal.GDP.Index*w.y,
                            Terciarias=Monthly.Nominal.GDP.Index*w) %>% 
  dplyr::select(OBS, Total, Primarias, Secundarias, Terciarias)


# Year percentage changes
GDP_US[, -1] <- sapply(GDP_US[, -1], pch)
GDP_US <- GDP_US %>% filter(!is.na(Total))

# Saving
GDP_US %>% write.csv("Data/GDP_US_Clean.csv", row.names = F)

```
